% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipcancer.R
\docType{data}
\name{bear}
\alias{bear}
\title{Bear dataset }
\description{
Number of physiological measurements from 54 bears.
}
\format{
data: A data frame with 54 rows and the following 9 variables:
\describe{
  \item{age}{}
  \item{length}{}
  \item{sex}{}
  \item{weight}{}
  \item{chest}{}
  \item{headlength}{}
  \item{headwid}{}
  \item{month}{}
  \item{neck}{}
}

}
\keyword{datasets}
