% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipcancer.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{Ozone data}
\description{
data set consists of 112 measurements of maximum daily ozone in
Rennes.  In addition,  temperature (T), nebulosity (Ne), and projection of
wind speed vectors (Vx) were measured three times daily (9:00, 12:00, and
15:00 hours) resulting in nine covariates.
}
\format{
data: A data frame with 112 rows and the following variables:
\describe{
  \item{num}{observed number of cancer cases}
  \item{maxO3}{max daily ozone}
  \item{T9-T15}{temperature measured at 9:00, 12:00, and 15:00 hours}
  \item{Ne9-Ne15}{nebulosity measured at 9:00, 12:00, and 15:00 hours}
  \item{Vx9-Vx15}{projection of wind speed vectors  measured at 9:00, 12:00, and 15:00 hours}
  \item{max03v}{max daily ozone of previous day.}
  \item{WindDirection}{The wind direction}
}
}
\source{
Source of data: \url{https://github.com/njtierney/user2018-missing-data-tutorial/}
}
