% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipcancer.R
\docType{data}
\name{scallops}
\alias{scallops}
\title{Scallops data}
\description{
Data set that provides the location and scallop catches in the Atlantic waters off the coasts of New Jersey and Long Island, New York.
}
\format{
data:  A data frame with 148 rows and the variables are the following:
\describe{
  \item{strata}{}
  \item{sample}{}
  \item{lat}{}
  \item{long}{}
  \item{tcatch}{}
  \item{prerec}{}
  \item{recruits}{}
}

}
\source{
Banerjee, S; Carline, B. P.; Gelfand, A. E.; (2015) Hierarchical Modeling and Analysis for Spatial Data 2nd Ed. CRC. Press
}
\keyword{datasets}
