% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipcancer.R
\docType{data}
\name{SIMCE}
\alias{SIMCE}
\title{Standardized testing data in Chile}
\description{
Average standard testing results with average mother's and father's education level for schools in the greater Santiago area of Chile.  Measurements are recorded from 2005-2011 and spatial coordinates of
the schools are provided.
}
\format{
data: A data frame with 1072 rows and the following variables:
\describe{
  \item{coords.x1}{longitude coordinates of school}
  \item{coords.x2}{lattitude coordinates of school}
  \item{Schoole}{Unique school identifier}
  \item{COMUNA}{Name of the commune in which the school resides}
  \item{SIMCE05-SIMCE11}{Math score of standardized test in 2005-2011}
  \item{EDpad05-EDpad11}{Average level of father's education of students that attended school 2005-2011}
  \item{EDmad05-EDmad11}{Average level of mother's education of students that attended school 2005-2011}
}
}
\source{
Page, G. L. and Quintana, F. A. (2016) Spatial Product Partition Models, Bayesian Anal., Volume 11, Number 1, 265-298.
}
